Refactor message handling
